rbac_v1/requirements.txt
asgiref==3.5.2
Django==3.2.13
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
djangorestframework==3.13.1
PyJWT==2.4.0
django-filter==21.1